site/scripts/requirements.txt
sentence-transformers>=2.2.2
numpy>=1.21.0
scikit-learn>=1.0.2
mysql-connector-python>=8.0.0
